v2.0
Improved code modularity. If you want to switch to a different type of displays or sensors, it will be easier for you to modify something to adapt it to your own needs.
Some additions, changes and fixes. Cleaned up the repeated code. Refactoring for better code understanding. More comments added.
Now it works with OWM_for_ESP v2.0.0